Tech/gradio/requirements.txt
gradio>=3.50
requests>=2.31.0
python-telegram-bot>=20.3
python-dotenv>=1.0.0
